Correcciones a variables

+++ Pinout.h
@@ -8,15 +8,15 @@
 // I2C_LCD_SCL = A5;
 
 // Driver Motores L298N
-const byte ENB_PIN = 5;
-const byte IN3_PIN = 9;
-const byte IN4_PIN = 10;
-const byte ENA_PIN = 6;
-const byte IN1_PIN = 7;
-const byte IN2_PIN = 8;
+const int ENB_PIN = 5;
+const int IN3_PIN = 9;
+const int IN4_PIN = 10;
+const int ENA_PIN = 6;
+const int IN1_PIN = 7;
+const int IN2_PIN = 8;
 
 // Sensores de línea
-const byte SENSOR_LL_PIN = 4;
-const byte SENSOR_L_PIN = 11;
-const byte SENSOR_RR_PIN = 13;
-const byte SENSOR_R_PIN = 12;
+const int SENSOR_LL_PIN = 4;
+const int SENSOR_L_PIN = 11;
+const int SENSOR_RR_PIN = 13;
+const int SENSOR_R_PIN = 12;
